pyproject.toml
[tool.poetry]
name = "psycho-support"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standart"], version = "^0.29.0"}
SQLAlchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
cryptography = "^42.0.5"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
